# Use a freelist for struct types