@jcrist 2021-07-09
Running the benchmark script gives:

**master**

```
- msgspec...
  dumps: 224.36 us
  loads: 688.29 us
- msgspec asarray...
  dumps: 154.70 us
  loads: 453.03 us
```

**This PR**

```
- msgspec...
  dumps: 222.62 us
  loads: 622.18 us
- msgspec asarray...
  dumps: 152.51 us
  loads: 435.89 us
```

For 10% and 4% speedups on load time respectively.
```